Unicode WIP, does not compile. - finishing for today. Spells still need work, the error-messages need translation.

+++ src/common/gamecode/randenc.c
@@ -245,7 +245,7 @@ find_manual(region * r, unit * u)
     break;
   }
 
-  sprintf(zLocation, "manual_location_%d", rng_int() % 4);
+  sprintf(zLocation, "manual_location_%d", (int)(rng_int() % 4));
   sprintf(zBook, "manual_title_%s", skillnames[skill]);
 
   msg = msg_message("find_manual", "unit location book", u, zLocation, zBook);
